Remove unnecessary Primitives wrapper
